// src/app/domain/dto/evaluate-plans-tree-dto.ts
export interface EvaluatePlansTreeDTO {
    entidadeId: number;
    entidadeNome: string;
    cicloId: number;
    cicloNome: string;
    cicloNota: number;
    pilarId: number;
    pilarNome: string;
    pilarPeso: number;
    pilarNota: number;
    componenteId: number;
    componenteNome: string;
    componentePeso: number;
    componenteNota: number;
    supervisorId: number;
    supervisorNome: string;
    auditorId: number;
    tipoNotaId: number;
    tipoNotaLetra: string;
    tipoNotaCorLetra: string;
    tipoNotaNome: string;
    tipoNotaPeso: number;
    tipoNotaNota: number;
    elementoId: number;
    elementoNome: string;
    elementoPeso: number;
    elementoNota: number;
    tipoPontuacaoId: number;
    pesoPadraoEC: number;
    tipoMediaCPId: number;
    pesoPadraoCP: number;
    tipoMediaPCId: number;
    pesoPadraoPC: number;
    itemId: number;
    itemNome: string;
    planoId: number;
    cnpb: string;
    recursoGarantidor: string;
    planoModalidade: number;
    planoPeso: number;
    planoNota: number;
    iniciaEm: string;
    terminaEm: string;
    cStatus: string;
    statusId: number;
    produtoComponenteId: number;
    periodoPermitido: number;
    periodoCiclo: number;
}